refactor(game-sessions): consolidate type guards and update import paths for better organization
