+++ packages/shared/src/utils/index.ts
@@ -1,2 +1,2 @@
 export * from "./date"
-export * from "./typeGuards"
+export * from "./type-guards"
